chore: extract types into a shared libary for engine-controller canister
