+++ rs/engine_controller/BUILD.bazel
@@ -1,7 +1,19 @@
-load("@rules_rust//rust:defs.bzl", "rust_test")
+load("@rules_rust//rust:defs.bzl", "rust_library", "rust_test")
 load("//bazel:canisters.bzl", "rust_canister")
 
 package(default_visibility = ["//visibility:public"])
+
+rust_library(
+    name = "ic_engine_controller",
+    srcs = ["src/lib.rs"],
+    crate_name = "ic_engine_controller",
+    deps = [
+        # Keep sorted.
+        "//rs/registry/canister",
+        "@crate_index//:candid",
+        "@crate_index//:serde",
+    ],
+)
 
 rust_canister(
     name = "engine-controller-canister",
@@ -10,6 +22,7 @@ rust_canister(
     service_file = ":engine_controller.did",
     deps = [
         # Keep sorted.
+        ":ic_engine_controller",
         "//rs/nns/constants",
         "//rs/protobuf",
         "//rs/registry/canister",
@@ -17,7 +30,6 @@ rust_canister(
         "//rs/types/base_types",
         "@crate_index//:candid",
         "@crate_index//:ic-cdk",
-        "@crate_index//:serde",
     ],
 )
 
@@ -28,6 +40,7 @@ rust_test(
     crate_root = "canister/canister.rs",
     deps = [
         # Keep sorted.
+        ":ic_engine_controller",
         "//rs/nns/constants",
         "//rs/protobuf",
         "//rs/registry/canister",
@@ -36,7 +49,6 @@ rust_test(
         "@crate_index//:candid",
         "@crate_index//:candid_parser",
         "@crate_index//:ic-cdk",
-        "@crate_index//:serde",
     ],
 )
 
@@ -59,6 +71,7 @@ rust_test(
     },
     deps = [
         # Keep sorted.
+        ":ic_engine_controller",
         "//packages/pocket-ic",
         "//rs/nns/constants",
         "//rs/nns/test_utils",
@@ -72,7 +85,6 @@ rust_test(
         "@crate_index//:candid",
         "@crate_index//:ic-management-canister-types",
         "@crate_index//:prost",
-        "@crate_index//:serde",
         "@crate_index//:tokio",
     ],
 )
